Adding conspect controller and making some minor changes
