+++ NOTEA/Views/Conspect/ConspectList.cshtml
@@ -1,5 +1,4 @@
-﻿@* @model NOTEA.Models.ConspectModel *@
-@model NOTEA.Models.ConspectListModel
+﻿@model NOTEA.Models.ConspectListModel
  
 @{
     ViewData["Title"] = "Conspect list";
@@ -19,7 +18,6 @@
                     @conspectModel.Date
                 </td>
             </tr>
-            @* <span>@conspectModel.Name </span> *@
         }
     </tbody>
 </table>
