--- NOTEA/Views/Home/HowToUse.cshtml
@@ -1,10 +1,0 @@
-﻿
-@{
-    ViewData["Title"] = "How to use";
-}
-
-<h1>How to use this app:</h1>
-<p>1. Click on "Create Conspects"</p>
-<p>2. Start writing</p>
-<p>3. Press "Save"</p>
-
