+++ NOTEA/Views/Shared/_Layout.cshtml
@@ -26,16 +26,13 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="HowToUse">How to use</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Conspect" asp-action="CreateConspects">Create conspects</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="CreateConspects">Create conspects</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Conspect" asp-action="ConspectList">List of conspects</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="ConspectList">List of conspects</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="UploadConspect">Upload conspects</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="Conspect" asp-action="UploadConspect">Upload conspects</a>
                         </li>
                     </ul>
                 </div>
